refactor(Lotto):Lotto 클래스 도메인 패키지 하위로 이동 및 소스코드 정리
