+++ src/main/java/lotto/domain/Lotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -20,47 +20,32 @@ public class Lotto {
     }
 
     private void validate(List<Integer> numbers) {
-        if (numbers.size() != LOTTO_SIZE) {
+        sizeCheck(numbers);
+        duplicateCheck(numbers);
+        limitCheck(numbers);
+    }
+
+    private void sizeCheck(List<Integer> numbers) {
+        if(numbers.size() != LOTTO_SIZE) {
             throw new IllegalArgumentException(SIZE_CHECK_ERROR);
-        }
-        if (duplicateCheck(numbers)) {
-            throw new IllegalArgumentException(DUPLICATE_CHECK_ERROR);
-        }
-        if (limitCheck(numbers)) {
-            throw new IllegalArgumentException(LIMIT_CHECK_ERROR);
         }
     }
 
-    // TODO: 추가 기능 구현
-    private boolean duplicateCheck(List<Integer> numbers) {
+    private void duplicateCheck(List<Integer> numbers) {
         List<Integer> tempNumbers = new ArrayList<>();
         for(int i = 0; i < numbers.size(); i++) {
             if(tempNumbers.contains(numbers.get(i))) {
-                return true;
+                throw new IllegalArgumentException(DUPLICATE_CHECK_ERROR);
             }
             tempNumbers.add(numbers.get(i));
         }
-        return false;
     }
-    private boolean limitCheck(List<Integer> numbers) {
+    private void limitCheck(List<Integer> numbers) {
         for(int i = 0; i < numbers.size(); i++) {
             if(numbers.get(i) > LOTTO_END || numbers.get(i) < LOTTO_START) {
-                return true;
+                throw new IllegalArgumentException(LIMIT_CHECK_ERROR);
             }
         }
-        return false;
     }
-
-    private void buyLotto(int price) {
-        if(price % 1000 != 0) {
-            throw new IllegalArgumentException("[ERROR] 구입 금액이 1,000원 단위가 아닙니다.");
-        }
-        for (int i = 0; i < price/1000; i++) {
-            Randoms.pickUniqueNumbersInRange(1, 45, 6);
-        }
-
-    }
-
-
 
 }
