+++ src/main/java/lotto/view/InputView.java
@@ -1,0 +1,10 @@
+package lotto.view;
+
+import camp.nextstep.edu.missionutils.Console;
+
+public class Input {
+    private static final String PURCHASE_LOTTO_INPUT_MESSAGE = "구입금액을 입력해주세요.";
+    private static final String PURCHASE_LOTTO_COUNT_MESSAGE = "당첨 번호를 입력후재세요"
+
+
+}
